refactor: move vc-related functions to the version control context
