+++ src/modules/editor-state/current-document/context.tsx
@@ -1,4 +1,3 @@
-import debounce from 'debounce';
 import {
   createContext,
   useCallback,
@@ -14,22 +13,15 @@ import { FilesystemContext } from '../../filesystem/react';
 import { FunctionalityConfigContext } from '../../personalization/functionality-config';
 import {
   type Change,
-  type ChangeWithUrlInfo,
-  type Commit,
   convertToStorageFormat,
   type DocHandleChangePayload,
   encodeURLHeads,
   encodeURLHeadsForChange,
-  getDocumentHandleHistory,
-  getDocumentHeads,
   headsAreSame,
-  isContentSameAtHeads,
-  isEmpty,
   isValidVersionControlId,
   type RichTextDocument,
   type UrlHeads,
   type VersionControlId,
-  VersionedDocument,
   type VersionedDocumentHandle,
 } from '../../version-control';
 import { VersionControlContext } from '../../version-control/react';
@@ -44,13 +36,6 @@ type CurrentDocumentContextType = {
   selectedFileName: string | null;
   setSelectedFileInfo: (file: SelectedFileInfo) => Promise<void>;
   clearFileSelection: () => Promise<void>;
-  versionedDocumentHandle: VersionedDocumentHandle | null;
-  versionedDocumentHistory: ChangeWithUrlInfo[];
-  canCommit: boolean;
-  onCommit: (message: string) => void;
-  isCommitDialogOpen: boolean;
-  onOpenCommitDialog: () => void;
-  onCloseCommitDialog: () => void;
   selectedCommitIndex: number | null;
   onSelectCommit: (heads: UrlHeads) => void;
 };
@@ -61,13 +46,6 @@ export const CurrentDocumentContext = createContext<CurrentDocumentContextType>(
     selectedFileName: null,
     setSelectedFileInfo: async () => {},
     clearFileSelection: async () => {},
-    versionedDocumentHandle: null,
-    versionedDocumentHistory: [],
-    canCommit: false,
-    onCommit: () => {},
-    isCommitDialogOpen: false,
-    onOpenCommitDialog: () => {},
-    onCloseCommitDialog: () => {},
     selectedCommitIndex: null,
     onSelectCommit: () => {},
   }
@@ -82,18 +60,16 @@ export const CurrentDocumentProvider = ({
   const [selectedFileInfo, setSelectedFileInfo] =
     useState<SelectedFileInfo | null>(null);
   const [selectedFileName, setSelectedFileName] = useState<string | null>(null);
-  const [versionedDocumentHandle, setVersionedDocumentHandle] =
-    useState<VersionedDocumentHandle | null>(null);
   const { documentId } = useParams();
   const [searchParams] = useSearchParams();
-  const { findDocument } = useContext(VersionControlContext);
+  const {
+    findDocument,
+    versionedDocumentHandle,
+    setVersionedDocumentHandle,
+    versionedDocumentHistory,
+  } = useContext(VersionControlContext);
   const { writeFile } = useContext(FilesystemContext);
-  const [versionedDocumentHistory, setVersionedDocumentHistory] = useState<
-    ChangeWithUrlInfo[]
-  >([]);
-  const [lastCommit, setLastCommit] = useState<Commit | null>(null);
-  const [canCommit, setCanCommit] = useState(false);
-  const [isCommitDialogOpen, setIsCommitDialogOpen] = useState<boolean>(false);
+
   const [selectedCommitIndex, setSelectedCommitIndex] = useState<number | null>(
     null
   );
@@ -150,55 +126,6 @@ export const CurrentDocumentProvider = ({
     }
   }, [selectedFileInfo]);
 
-  const checkIfContentChangedFromLastCommit = (
-    currentDoc: VersionedDocument,
-    latestChangeHeads: UrlHeads,
-    lastCommitHeads: UrlHeads
-  ) => {
-    if (
-      !headsAreSame(latestChangeHeads, lastCommitHeads) &&
-      !isContentSameAtHeads(currentDoc, latestChangeHeads, lastCommitHeads)
-    ) {
-      setCanCommit(true);
-    } else {
-      setCanCommit(false);
-    }
-  };
-
-  const checkIfCanCommit = (
-    currentDoc: VersionedDocument,
-    latestChangeHeads: UrlHeads,
-    lastCommitHeads?: UrlHeads
-  ) => {
-    if (lastCommitHeads) {
-      checkIfContentChangedFromLastCommit(
-        currentDoc,
-        latestChangeHeads,
-        lastCommitHeads
-      );
-    } else {
-      if (!isEmpty(currentDoc)) {
-        setCanCommit(true);
-      } else {
-        setCanCommit(false);
-      }
-    }
-  };
-
-  const loadHistory = async (docHandle: VersionedDocumentHandle) => {
-    const { history, currentDoc, lastCommit, latestChange } =
-      await getDocumentHandleHistory(docHandle);
-
-    const historyWithURLInfo = history.map((commit) => ({
-      ...commit,
-      urlEncodedHeads: encodeURLHeadsForChange(commit),
-    }));
-
-    setVersionedDocumentHistory(historyWithURLInfo);
-    setLastCommit(lastCommit);
-    checkIfCanCommit(currentDoc, latestChange.heads, lastCommit?.heads);
-  };
-
   useEffect(() => {
     const updateDocTitle = async (docHandle: VersionedDocumentHandle) => {
       const doc = await docHandle.doc();
@@ -207,30 +134,8 @@ export const CurrentDocumentProvider = ({
 
     if (versionedDocumentHandle) {
       updateDocTitle(versionedDocumentHandle);
-      loadHistory(versionedDocumentHandle);
     }
   }, [versionedDocumentHandle]);
-
-  useEffect(() => {
-    if (versionedDocumentHandle) {
-      const handler = (args: DocHandleChangePayload<VersionedDocument>) => {
-        loadHistory(versionedDocumentHandle);
-        checkIfCanCommit(
-          args.doc,
-          getDocumentHeads(args.doc),
-          lastCommit?.heads
-        );
-      };
-
-      const debouncedHandler = debounce(handler, 300);
-      versionedDocumentHandle.on('change', debouncedHandler);
-
-      return () => {
-        versionedDocumentHandle.off('change', debouncedHandler);
-      };
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [lastCommit, versionedDocumentHandle]);
 
   const clearFileSelection = async () => {
     setSelectedFileInfo(null);
@@ -249,38 +154,6 @@ export const CurrentDocumentProvider = ({
       path: path,
     });
   };
-
-  const commitChanges = useCallback(
-    (message: string) => {
-      if (!versionedDocumentHandle) return;
-
-      versionedDocumentHandle.change(
-        (doc) => {
-          // this is effectively a no-op, but it triggers a change event
-          // (not) changing the title of the document, as interfering with the
-          // content outside the Prosemirror API will cause loss of formatting
-          // eslint-disable-next-line no-self-assign
-          doc.title = doc.title;
-        },
-        {
-          message,
-          time: new Date().getTime(),
-        }
-      );
-
-      setIsCommitDialogOpen(false);
-      setCanCommit(false);
-    },
-    [versionedDocumentHandle]
-  );
-
-  const handleOpenCommitDialog = useCallback(() => {
-    setIsCommitDialogOpen(true);
-  }, []);
-
-  const handleCloseCommitDialog = useCallback(() => {
-    setIsCommitDialogOpen(false);
-  }, []);
 
   const handleSelectCommit = useCallback(
     (heads: UrlHeads) => {
@@ -322,15 +195,8 @@ export const CurrentDocumentProvider = ({
       value={{
         selectedFileInfo,
         selectedFileName,
-        versionedDocumentHandle,
         setSelectedFileInfo: handleSetSelectedFileInfo,
         clearFileSelection,
-        versionedDocumentHistory,
-        canCommit,
-        onCommit: commitChanges,
-        isCommitDialogOpen,
-        onOpenCommitDialog: handleOpenCommitDialog,
-        onCloseCommitDialog: handleCloseCommitDialog,
         selectedCommitIndex,
         onSelectCommit: handleSelectCommit,
       }}
