+++ src/modules/version-control/react/context.tsx
@@ -1,5 +1,12 @@
+import debounce from 'debounce';
 import * as Effect from 'effect/Effect';
-import { createContext, useContext, useEffect, useState } from 'react';
+import {
+  createContext,
+  useCallback,
+  useContext,
+  useEffect,
+  useState,
+} from 'react';
 
 import { ElectronContext } from '../../electron';
 import { createAdapter } from '../adapters/automerge';
@@ -7,7 +14,21 @@ import {
   setupForElectron as setupBrowserRepoForElectron,
   setupForWeb as setupBrowserRepoForWeb,
 } from '../automerge-repo/browser';
-import { VersionControlId, VersionedDocumentHandle } from '../models';
+import {
+  type ChangeWithUrlInfo,
+  type Commit,
+  type DocHandleChangePayload,
+  encodeURLHeadsForChange,
+  getDocumentHandleHistory,
+  getDocumentHeads,
+  headsAreSame,
+  isContentSameAtHeads,
+  isEmpty,
+  type UrlHeads,
+  type VersionControlId,
+  type VersionedDocument,
+  type VersionedDocumentHandle,
+} from '../models';
 import {
   type CreateDocumentArgs,
   type FindDocumentInProjectArgs,
@@ -18,6 +39,14 @@ import {
 type VersionControlContextType = {
   versionControlRepo: VersionControlRepo | null;
   isRepoReady: boolean;
+  versionedDocumentHandle: VersionedDocumentHandle | null;
+  setVersionedDocumentHandle: (handle: VersionedDocumentHandle | null) => void;
+  versionedDocumentHistory: ChangeWithUrlInfo[];
+  canCommit: boolean;
+  onCommit: (message: string) => void;
+  isCommitDialogOpen: boolean;
+  onOpenCommitDialog: () => void;
+  onCloseCommitDialog: () => void;
   createDocument: (args: CreateDocumentArgs) => Promise<VersionControlId>;
   getDocumentHandleAtCommit: (
     args: GetDocumentHandleAtCommitArgs
@@ -33,6 +62,14 @@ type VersionControlContextType = {
 export const VersionControlContext = createContext<VersionControlContextType>({
   versionControlRepo: null,
   isRepoReady: false,
+  versionedDocumentHandle: null,
+  setVersionedDocumentHandle: () => {},
+  versionedDocumentHistory: [],
+  canCommit: false,
+  onCommit: () => {},
+  isCommitDialogOpen: false,
+  onOpenCommitDialog: () => {},
+  onCloseCommitDialog: () => {},
   // @ts-expect-error will get overriden below
   createDocument: () => null,
   // @ts-expect-error will get overriden below
@@ -52,6 +89,14 @@ export const VersionControlProvider = ({
     useState<VersionControlRepo | null>(null);
   const [isRepoReady, setIsRepoReady] = useState<boolean>(false);
   const { processId, isElectron } = useContext(ElectronContext);
+  const [versionedDocumentHandle, setVersionedDocumentHandle] =
+    useState<VersionedDocumentHandle | null>(null);
+  const [versionedDocumentHistory, setVersionedDocumentHistory] = useState<
+    ChangeWithUrlInfo[]
+  >([]);
+  const [lastCommit, setLastCommit] = useState<Commit | null>(null);
+  const [canCommit, setCanCommit] = useState(false);
+  const [isCommitDialogOpen, setIsCommitDialogOpen] = useState<boolean>(false);
 
   useEffect(() => {
     const setupVersionControlRepo = async () => {
@@ -74,6 +119,121 @@ export const VersionControlProvider = ({
 
     setupVersionControlRepo();
   }, [processId, isElectron]);
+
+  const checkIfCanCommit = (
+    currentDoc: VersionedDocument,
+    latestChangeHeads: UrlHeads,
+    lastCommitHeads?: UrlHeads
+  ) => {
+    if (lastCommitHeads) {
+      checkIfContentChangedFromLastCommit(
+        currentDoc,
+        latestChangeHeads,
+        lastCommitHeads
+      );
+    } else {
+      if (!isEmpty(currentDoc)) {
+        setCanCommit(true);
+      } else {
+        setCanCommit(false);
+      }
+    }
+  };
+
+  const checkIfContentChangedFromLastCommit = (
+    currentDoc: VersionedDocument,
+    latestChangeHeads: UrlHeads,
+    lastCommitHeads: UrlHeads
+  ) => {
+    if (
+      !headsAreSame(latestChangeHeads, lastCommitHeads) &&
+      !isContentSameAtHeads(currentDoc, latestChangeHeads, lastCommitHeads)
+    ) {
+      setCanCommit(true);
+    } else {
+      setCanCommit(false);
+    }
+  };
+
+  const loadHistory = async (docHandle: VersionedDocumentHandle) => {
+    const { history, currentDoc, lastCommit, latestChange } =
+      await getDocumentHandleHistory(docHandle);
+
+    const historyWithURLInfo = history.map((commit) => ({
+      ...commit,
+      urlEncodedHeads: encodeURLHeadsForChange(commit),
+    }));
+
+    setVersionedDocumentHistory(historyWithURLInfo);
+    setLastCommit(lastCommit);
+    checkIfCanCommit(currentDoc, latestChange.heads, lastCommit?.heads);
+  };
+
+  useEffect(() => {
+    if (versionedDocumentHandle) {
+      const handler = (args: DocHandleChangePayload<VersionedDocument>) => {
+        loadHistory(versionedDocumentHandle);
+        checkIfCanCommit(
+          args.doc,
+          getDocumentHeads(args.doc),
+          lastCommit?.heads
+        );
+      };
+
+      const debouncedHandler = debounce(handler, 300);
+      versionedDocumentHandle.on('change', debouncedHandler);
+
+      return () => {
+        versionedDocumentHandle.off('change', debouncedHandler);
+      };
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [lastCommit, versionedDocumentHandle]);
+
+  const commitChanges = useCallback(
+    (message: string) => {
+      if (!versionedDocumentHandle) return;
+
+      versionedDocumentHandle.change(
+        (doc) => {
+          // this is effectively a no-op, but it triggers a change event
+          // (not) changing the title of the document, as interfering with the
+          // content outside the Prosemirror API will cause loss of formatting
+          // eslint-disable-next-line no-self-assign
+          doc.title = doc.title;
+        },
+        {
+          message,
+          time: new Date().getTime(),
+        }
+      );
+
+      setIsCommitDialogOpen(false);
+      setCanCommit(false);
+    },
+    [versionedDocumentHandle]
+  );
+
+  useEffect(() => {
+    if (versionedDocumentHandle) {
+      loadHistory(versionedDocumentHandle);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [versionedDocumentHandle]);
+
+  const handleOpenCommitDialog = useCallback(() => {
+    setIsCommitDialogOpen(true);
+  }, []);
+
+  const handleCloseCommitDialog = useCallback(() => {
+    setIsCommitDialogOpen(false);
+  }, []);
+
+  const handleSetVersionedDocumentHandle = (
+    handle: VersionedDocumentHandle | null
+  ) => {
+    setVersionedDocumentHandle(handle);
+  };
 
   const handleCreateDocument = async (args: CreateDocumentArgs) => {
     if (!versionControlRepo) {
@@ -121,6 +281,14 @@ export const VersionControlProvider = ({
       value={{
         versionControlRepo,
         isRepoReady,
+        versionedDocumentHandle,
+        setVersionedDocumentHandle: handleSetVersionedDocumentHandle,
+        versionedDocumentHistory,
+        canCommit,
+        onCommit: commitChanges,
+        isCommitDialogOpen,
+        onOpenCommitDialog: handleOpenCommitDialog,
+        onCloseCommitDialog: handleCloseCommitDialog,
         createDocument: handleCreateDocument,
         getDocumentHandleAtCommit: handleGetDocumentHandleAtCommit,
         findDocument: handleFindDocument,
