+++ src/renderer/src/pages/document/main/editor/DocumentEditor.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useContext, useState } from 'react';
 
-import { CurrentDocumentContext } from '../../../../../../modules/editor-state';
 import { SidebarLayoutContext } from '../../../../../../modules/editor-state/sidebar-layout/context';
 import { ProseMirrorContext } from '../../../../../../modules/rich-text/react/context';
+import { VersionControlContext } from '../../../../../../modules/version-control/react';
 import { RichTextEditor } from '../../../../components/editing/RichTextEditor';
 import { ActionsBar } from './ActionsBar';
 
@@ -10,7 +10,7 @@ export const DocumentEditor = () => {
   const [isEditorToolbarOpen, toggleEditorToolbar] = useState<boolean>(false);
   const { view: editorView } = useContext(ProseMirrorContext);
   const { versionedDocumentHandle, onOpenCommitDialog, canCommit } = useContext(
-    CurrentDocumentContext
+    VersionControlContext
   );
   const { isSidebarOpen, toggleSidebar } = useContext(SidebarLayoutContext);
